fixed compilation error

+++ 09/polynomial.hpp
@@ -17,7 +17,7 @@ namespace leic {
       return coeffs.size() - 1;
     }
     void add(const polynomial& p) {
-      int n = std::min(coeffs.size(), p.coeffs.size());
+      size_t n = std::min(coeffs.size(), p.coeffs.size());
       for (size_t i = 0; i < n; i++) {
         coeffs[i].add(p.coeffs[i]);
       }
